Added database connection
Database is now connected to the application. Queries can run and dynamically update parts of the home page. Current issue is that the project roles list is loaded asynchronously which can cause it to lag behind the webpage loading and the omission of some values. Adding in some callbacks to fix this in the future.